requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
astor==0.8.1
attrs==25.3.0
cachetools==5.5.2
certifi==2025.6.15
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
colorlog==6.9.0
cssselect==1.3.0
cssutils==2.11.1
dataclasses-json==0.6.7
decorator==5.2.1
distro==1.9.0
dnspython==2.7.0
einops==0.8.1
email_validator==2.2.0
et_xmlfile==2.0.0
faiss-cpu==1.11.0
fastapi==0.115.13
fastapi-cli==0.0.7
filelock==3.18.0
filetype==1.2.0
frozenlist==1.7.0
fsspec==2025.5.1
ftfy==6.3.1
google-ai-generativelanguage==0.6.18
google-api-core==2.25.1
google-auth==2.40.3
googleapis-common-protos==1.70.0
GPUtil==1.4.0
grpcio==1.73.0
grpcio-status==1.73.0
h11==0.16.0
hf-xet==1.1.4
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.33.0
idna==3.10
imagesize==1.4.1
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.25
langchain-community==0.3.25
langchain-core==0.3.65
langchain-google-genai==2.1.5
langchain-openai==0.3.24
langchain-text-splitters==0.3.8
langsmith==0.3.45
lxml==5.4.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mdurl==0.1.2
more-itertools==10.7.0
multidict==6.5.0
mypy_extensions==1.1.0
networkx==3.5
numpy==2.3.0
openai==1.88.0
opencv-contrib-python==4.10.0.84
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
opt-einsum==3.3.0
orjson==3.10.18
packaging==24.2
paddleocr==3.0.2
paddlepaddle==3.0.0
paddlex==3.0.2
pandas==2.3.0
pillow==11.2.1
premailer==3.10.0
prettytable==3.16.0
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.31.1
py-cpuinfo==9.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyclipper==1.3.0.post6
pydantic==2.11.7
pydantic-extra-types==2.10.5
pydantic-settings==2.9.1
pydantic_core==2.33.2
Pygments==2.19.1
pypdfium2==4.30.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
requests-toolbelt==1.0.0
rich==14.0.0
rich-toolkit==0.14.7
rsa==4.9.1
ruamel.yaml==0.18.14
ruamel.yaml.clib==0.2.12
scikit-learn==1.7.0
scipy==1.15.3
setuptools==80.9.0
shapely==2.1.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.1
tqdm==4.67.1
typer==0.16.0
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
ujson==5.10.0
urllib3==2.5.0
uvicorn==0.34.3
uvloop==0.21.0
watchfiles==1.1.0
wcwidth==0.2.13
websockets==15.0.1
yarl==1.20.1
zstandard==0.23.0
